requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
python-socketio==5.10.0
python-multipart==0.0.6
websockets==12.0
opencv-python
albumentations